+++ backend/graph_normalizer/graph_normalizer.py
@@ -3,7 +3,7 @@ import os
 from collectors.collect_metadata import MetadataFileRecord
 from constants import FileType
 from table_inventory.table_inventory import TableInventoryResult
-from utils import format_node_info
+from graph_normalizer.utils import format_node_info
 
 
 class GraphNormalizer:
